Update Video tests
