+++ .changeset/wise-masks-shop.md
@@ -1,0 +1,5 @@
+---
+'@shopify/hydrogen': patch
+---
+
+Allow `sourceProps` to be passed to `<Video/>`'s underlying `<source>` elements.
