+++ package.json
@@ -19,7 +19,7 @@
     "lint:schema": "jsonlint --quiet packages/hydrogen/graphql.schema.json",
     "format": "prettier --write --ignore-unknown .",
     "format:check": "prettier --check --ignore-unknown .",
-    "test": "run-p test:base test:hydrogen-ui test:hydrogen:vite",
+    "test": "run-p test:base test:hydrogen:vite",
     "test:base": "jest",
     "test:coverage": "jest --coverage",
     "test-e2e": "jest --config jest-e2e.config.ts --runInBand",
